--- gtk-rust-host/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "gtk-rust-host"
-version = "0.1.0"
-edition = "2018"
-default-run = "host"
-
-[dependencies]
-exec = "*"
-fork = "*"
-glib = "*"
-gtk = { version = "*", package = "gtk4" }
-libc = "*"
-rand = "*"
-wasmi = "*"
-wasmer-runtime = "*"
-
-[lib]
-name = "common"
-path = "src/common.rs"
-
-[[bin]]
-name = "container-wasmer"
-path = "src/bin/container-wasmer.rs"
-
-[[bin]]
-name = "container-wasmi"
-path = "src/bin/container-wasmi.rs"
-
-[[bin]]
-name = "host"
-path = "src/bin/host.rs"
